# src/backend/static/swagger.yaml
definitions:
  CommonError:
    properties:
      message:
        description: |-
          Message of the error
          in: string
        type: string
        x-go-name: Message
      status:
        $ref: '#/definitions/FrameworkErrorCode'
    type: object
    x-go-package: communication-service/pkg/models/errors
  CommonResponse:
    properties:
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
    type: object
    x-go-package: communication-service/pkg/models/common
  FrameworkErrorCode:
    format: int16
    type: integer
    x-go-package: communication-service/pkg/models/errors
  SendPushNotificationToTokenCommand:
    properties:
      body:
        description: |-
          body of the notification
          in: body
        type: string
        x-go-name: Body
      title:
        description: |-
          title of the notification
          in: body
        type: string
        x-go-name: Title
      token:
        description: |-
          receiver's registered token
          in: body
        type: string
        x-go-name: Token
    required:
    - title
    - body
    - token
    type: object
    x-go-package: communication-service/pkg/application/push_notification/commands
  SendPushNotificationToTopicCommand:
    properties:
      body:
        description: |-
          body of the notification
          in: body
        type: string
        x-go-name: Body
      title:
        description: |-
          title of the notification
          in: body
        type: string
        x-go-name: Title
      topic:
        description: |-
          topic of notification
          in: body
        type: string
        x-go-name: Topic
    required:
    - title
    - body
    - topic
    type: object
    x-go-package: communication-service/pkg/application/push_notification/commands
host: 5100/
info:
  title: 'Communication Service:'
  version: 0.1.2
paths:
  /api/v1/notification/token/:
    post:
      consumes:
      - application/json
      description: send notification to a designated device
      operationId: SendPushNotificationToToken
      parameters:
      - in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/SendPushNotificationToTokenCommand'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: CommonResponse
          schema:
            $ref: '#/definitions/CommonResponse'
        "400":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - Notification
  /api/v1/notification/topic/:
    post:
      consumes:
      - application/json
      description: send notification to all subscribed devices to a specific topic
      operationId: SendPushNotificationToTopic
      parameters:
      - in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/SendPushNotificationToTopicCommand'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: CommonResponse
          schema:
            $ref: '#/definitions/CommonResponse'
        "400":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      tags:
      - Notification
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"